# docs/Eventix-API.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Eventix API",
    "description": "Eventix Event Booking System - Session-based auth, events, bookings, audit log. Import this file into Apifox or ApiDog for full API documentation.",
    "version": "1.1.0",
    "contact": {
      "name": "Eventix API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "tags": ["Eventix API / Events"],
        "summary": "List Events",
        "description": "List events with pagination. Optional search by name/description. Public: shows published and coming_soon. Authenticated admin (optional Bearer): shows all statuses including draft.",
        "operationId": "listEvents",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "Items per page (max 100)" },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search in name/description" },
          { "name": "sort_by", "in": "query", "schema": { "type": "string", "enum": ["created_at", "name", "remaining_spots"], "default": "created_at" }, "description": "Sort field" },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }, "description": "Sort order" }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Event" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": 1,
                      "name": "Summer Concert 2025",
                      "description": "Annual summer concert",
                      "capacity": 100,
                      "booked_count": 25,
                      "remaining_spots": 75,
                      "status": "published",
                      "created_at": "2025-03-01T10:00:00.000Z"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 10,
                    "total": 1,
                    "total_pages": 1,
                    "has_next": false,
                    "has_prev": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Eventix API / Events"],
        "summary": "Create Event",
        "description": "Create a new event. Requires admin role. Default status is draft.",
        "operationId": "createEvent",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "capacity"],
                "properties": {
                  "name": { "type": "string", "maxLength": 255, "example": "Summer Concert 2025" },
                  "description": { "type": "string", "nullable": true, "maxLength": 2000, "example": "Annual summer concert at the park" },
                  "capacity": { "type": "integer", "minimum": 1, "example": 100 },
                  "status": { "type": "string", "enum": ["draft", "coming_soon", "published", "cancelled", "completed"], "default": "draft" }
                }
              },
              "example": {
                "name": "Summer Concert 2025",
                "description": "Annual summer concert at the park",
                "capacity": 100,
                "status": "draft"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Event" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Authentication required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/events/{eventId}": {
      "get": {
        "tags": ["Eventix API / Events"],
        "summary": "Get Event by ID",
        "description": "Get a single event by ID. Public: only published and coming_soon. Authenticated admin (optional Bearer): all statuses.",
        "operationId": "getEventById",
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Event details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Event" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid event ID", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "error": { "code": "EVB404001", "message": "Event not found", "details": { "event_id": 999 } }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Eventix API / Events"],
        "summary": "Update Event",
        "description": "Update an event. Requires admin role. Capacity must be >= booked_count.",
        "operationId": "updateEvent",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "maxLength": 255 },
                  "description": { "type": "string", "nullable": true },
                  "capacity": { "type": "integer", "minimum": 1 },
                  "status": { "type": "string", "enum": ["draft", "coming_soon", "published", "cancelled", "completed"] }
                }
              },
              "example": {
                "name": "Summer Concert 2025 (Updated)",
                "description": "Updated description",
                "capacity": 120,
                "status": "published"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Event" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Authentication required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Event not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login or register"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "EVB400001" },
              "message": { "type": "string", "example": "Validation failed" },
              "details": { "type": "object", "nullable": true }
            }
          }
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "capacity": { "type": "integer" },
          "booked_count": { "type": "integer" },
          "remaining_spots": { "type": "integer" },
          "status": { "type": "string", "enum": ["draft", "coming_soon", "published", "cancelled", "completed"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "total_pages": { "type": "integer" },
          "has_next": { "type": "boolean" },
          "has_prev": { "type": "boolean" }
        }
      }
    }
  }
}